requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.88.1
aiohttp==3.9.1